% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_translate.R
\name{auto_translate}
\alias{auto_translate}
\alias{google_translate}
\alias{microsoft_translate}
\title{Auto translate messages}
\usage{
auto_translate(translation, overwrite = FALSE, api_key, engine = c("google",
  "microsoft"))

google_translate(translation, overwrite = FALSE,
  api_key = Sys.getenv("GOOGLE_TRANSLATE_API_KEY"))

microsoft_translate(translation, overwrite = FALSE,
  api_key = Sys.getenv("MICROSOFT_TRANSLATER_API_KEY"))
}
\arguments{
\item{translation}{An object of class \code{"po"} containing a message 
translation.}

\item{overwrite}{Logical. If \code{TRUE}, translations will be found for all
messages. If \code{FALSE}, only the messages that do not have an existing
translation will be translated.}

\item{api_key}{A string describing an API key for Google Translate or Microsoft 
Translator.}
}
\value{
An object of class \code{"po"} containing a message translation.
}
\description{
Automaticlly translate messages using Google Translate or Microsoft 
Translator.
}
\note{
To get a Google Translate API KEY, you must create a Google Cloud
account. See \url{https://cloud.google.com/translate/docs/getting-started}
to get started.
Likewise, to use Microsfot Translator, you need a Microsoft Azure account.
See \url{https://www.microsoft.com/en-us/translator/getstarted.aspx} to get 
started.#' @examples
\dontrun{
  # create example package
  pkg <- dummy_pkg()
  
  # setup pkg for localization
  use_localization(pkg)
  
  # generate Spanish translation in memory
  (tran <- make_translation("es", translator = "Some Person <example@example.com>"))
  
  auto_tran <- google_translate(tran)
  View(auto_tran$direct)
  View(auto_tran$countable)
}
}
