% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_gettext.R
\name{check_for_gettext}
\alias{check_for_gettext}
\title{Check for \samp{gettext} availability}
\usage{
check_for_gettext()
}
\value{
A logical.
}
\description{
A simple function to check whether \samp{gettext} is available on your system.
}
\note{
Windows binaries for \samp{gettext} can be installed from \url{http://www.stats.ox.ac.uk/pub/Rtools/goodies/gettext-tools.zip} or \url{https://github.com/mlocati/gettext-iconv-windows}.
}
\examples{
check_for_gettext()
}
\seealso{
\code{\link{use_localization}}
}

