\name{document_msgs}
\alias{document_msgs}
\alias{document_msgs}
\title{Create Message Documentation}
\description{Create Rd or Roxygen documentation for diagnostic messages}
\usage{
document_msgs(pkg = ".")
}
\arguments{
    \item{pkg}{The directory of an R package to extract diagnostic messages from. Path is passed to \code{\link[devtools]{as.package}}.}
}
\details{This function is invoked for its side-effect of creating either Rd or Roxygen documentation of package messages.}
%\value{}
\author{Thomas J. Leeper}
%\references{}
%\seealso{}
%\examples{}
