% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_translations.R
\name{edit_translation}
\alias{edit_translation}
\title{Interactive translation editing}
\usage{
edit_translation(translation)
}
\arguments{
\item{translation}{An object of class \code{"po"}, such as returned by \code{\link{make_translation}} or \code{\link{read_translation}}.}
}
\value{
\code{make_translation} and \code{read_translation} reutrn an object of class \code{"po"}.
}
\description{
Edit a translation object in-memory using a simple interface
}
\note{
Emacs users may find the gettext PO file editor more comfortable: \url{https://www.gnu.org/software/gettext/manual/gettext.html#PO-Mode}.
}
\examples{
\dontrun{
  # setup pkg for localization
  use_localization()
  
  # generate translation in memory
  (tran <- make_translation("es", translator = "Some Person <example@examle.com>"))
  
  # edit translations
  tran2 <- edit_translation(tran)
  # write to disk
  write_translation(tran2)
}
}
\author{
Thomas J. Leeper
}
\seealso{
\code{\link{make_translation}}
}

