% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_example_pkg.R
\name{extract_example_pkg}
\alias{extract_example_pkg}
\title{Extract the dummy package}
\usage{
extract_example_pkg(extract_to = tempdir())
}
\arguments{
\item{extract_to}{A string denoted the path to extract the zip file to.}
}
\value{
A string denoting a path to the root of the \code{translateme}
package. That is, the directroy returned by 
\code{file.path(extract_to, "translateme")}.
}
\description{
Extract the dummy \code{translateme} package from its zip file.
}
\examples{
extract_example_pkg()
}

