% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettext.R
\name{check_for_gettext}
\alias{check_for_gettext}
\alias{install_gettext}
\title{\samp{gettext} availability}
\usage{
check_for_gettext()

install_gettext(location)
}
\arguments{
\item{location}{A character string specifying a file path where the \samp{gettext} binaries should be installed.}
}
\value{
A logical.
}
\description{
Check whether \samp{gettext} is available on your system; and install for Windows.
}
\details{
\code{install_gettext} provides a Windows-only installer into the specified directory. That directory should be one included in \env{PATH} so that the binaries are available.
}
\note{
Windows binaries for \samp{gettext} can also be installed from \url{http://www.stats.ox.ac.uk/pub/Rtools/goodies/gettext-tools.zip} or \url{https://github.com/mlocati/gettext-iconv-windows}.
}
\examples{
\dontrun{
  install_gettext()
}
check_for_gettext()
}
\seealso{
\code{\link{use_localization}}
}

