% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_translations.R
\name{check_translations}
\alias{check_translations}
\alias{install_translations}
\title{Install translations}
\usage{
check_translations(pkg = ".")

install_translations(pkg = ".", check = TRUE)
}
\arguments{
\item{pkg}{The directory of an R package to extract diagnostic messages from. Path is passed to \code{\link[devtools]{as.package}}.}

\item{check}{A logical indicating whether to run \code{check_translations} before installing translations.}

\item{which}{Currently ignored.}
}
\description{
Check and install message translations
}
\details{
\code{check_translations} is just a wrapper around \code{\link[tools]{checkPoFiles}}
}
\examples{
\dontrun{
   # check translations
   check_translations()
   
   # install translations
   install_translations()
}
}

