% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGtools-package.R
\docType{package}
\name{MSGtools}
\alias{MSGtools}
\alias{MSGtools-package}
\title{Handling of MSG SEVIRI data}
\description{
R functions to handle MSG SEVIRI data that have been processed with the Marburg MSG processing routines
}
\details{
Handling of MSG SEVIRI data
}
\author{
Hanna Meyer, Johannes Drönner
\cr
\emph{Maintainer:} Hanna Meyer \email{hanna.meyer@uni-muenster.de}
}
