% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_directory.R
\name{make_po_dir}
\alias{make_po_dir}
\title{Utilities for translation directory}
\usage{
make_po_dir(pkg = ".", verbose = getOption("verbose"))
}
\arguments{
\item{pkg}{The directory of an R package to extract diagnostic messages from. Path is passed to \code{\link[devtools]{as.package}}.}

\item{verbose}{A logical. Should the function be chatty?}
}
\value{
The path to the directory.
}
\description{
Check for and create a /po directory
}
\examples{
# create example package
pkg <- dummy_pkg()
if (!check_for_po_dir(pkg)) {
  make_po_dir(pkg)
}
}

