% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setlang.R
\name{setlang}
\alias{getlang}
\alias{resetlang}
\alias{setlang}
\title{Set or reset locale language}
\usage{
setlang(language)

getlang()

resetlang(language)
}
\arguments{
\item{language}{A character string containing a language.}
}
\value{
Invisibly, a logical indicating with the language setting was successful.
}
\description{
Set or reset a locale language
}
\details{
\code{getlang} retrieves the current value of \samp{LANG} environment variables. \code{setlang} sets the \samp{LANG} environment variable and stores the supplied value of \code{language} in \code{options("msgtools_LANG")}. \code{resetlang} resets the language to the value specified in that option or, if that is missing, to English (\dQuote{EN}). This can be useful for testing message translations in various languages.
}
\author{
Thomas J. Leeper
}
\seealso{
\code{\link[poio]{language_codes}}
}

