% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_pkg.R
\name{dummy_pkg}
\alias{dummy_pkg}
\title{Dummy package creation}
\usage{
dummy_pkg(location = tempdir())
}
\arguments{
\item{location}{A character string containing a path to a directory in which to create the package, by default the location of \code{\link[base]{tempdir}}.}
}
\value{
A string denoting a path to the root of the package (called \dQuote{translateme}).
}
\description{
Create a simple example package in the temporary directory
}

