% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_messages.R
\name{get_messages}
\alias{get_messages}
\title{Extract diagnostic messages}
\usage{
get_messages(pkg = ".")
}
\arguments{
\item{pkg}{The directory of an R package to extract diagnostic messages from. Path is passed to \code{\link[devtools]{as.package}}.}
}
\value{
A \dQuote{tibble} (data frame) containing messages, pluralized messages, and the file location of each message.
}
\description{
Extracts diagnostic messages from a package using \code{\link[tools]{xgettext}}
}
\details{
Extracts diagnostic messages from a package.
}
\examples{
\dontrun{
  # get tidy data frame of messages
  get_messages()

  # spell check messages
  spell_check_msgs()
}
}
\author{
Thomas J. Leeper
}
\seealso{
\code{\link{spell_check_msgs}}
}

