\name{msgtools-package}
\alias{msgtools}
\alias{msgtools-package}
\docType{package}
\title{Tools for developing diagnostic messages}
\description{This package implements a number utilities for developing and maintaining error, warning, and other messages in R packages, including checking for consistency across messages, spell-checking messages, and building message translations into various languages.}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
