% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_directory.R
\name{check_for_po_dir}
\alias{check_for_po_dir}
\alias{make_po_dir}
\title{Utilities for translation directory}
\usage{
check_for_po_dir(pkg = ".")

make_po_dir(pkg = ".")
}
\arguments{
\item{pkg}{The directory of an R package to extract diagnostic messages from. Path is passed to \code{\link[devtools]{as.package}}.}
}
\value{
\code{check_for_po_dir} returns a logical. \code{make_po_dir} returns the path to the directory.
}
\description{
Check for and create a /po directory
}
\examples{
# from within a package directory
pkg <- extract_example_pkg()
if (!check_for_po_dir(pkg)) {
  make_po_dir(pkg)
}
}

