% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{make_template}
\alias{check_for_template}
\alias{make_template}
\alias{read_template}
\alias{sync_template}
\alias{template_current}
\alias{write_template}
\title{Handle message templates (.pot files)}
\usage{
make_template(charset = "UTF-8", pkg = ".", domain = "R")

sync_template(charset = "UTF-8", pkg = ".", domain = "R")

read_template(pkg = ".", domain = "R")

write_template(template, pkg = ".")

check_for_template(pkg = ".", domain = "R")

template_current(template, pkg = ".", domain = "R")
}
\arguments{
\item{charset}{A character string specifying the character set of the translation template file.}

\item{pkg}{The directory of an R package to extract diagnostic messages from. Path is passed to \code{\link[devtools]{as.package}}.}

\item{domain}{A character string specifying the \dQuote{domain} of the messages. Either \dQuote{R} (the default) or \dQuote{C}. This is case insensitive.}

\item{template}{An object of class \dQuote{po} containing a message translation template.}
}
\value{
\code{make_template} and \code{read_template} reutrn an object of class \dQuote{po}. \code{write_template} returns the path to the file, invisibly.
}
\description{
Read, write, and generate .pot diagnostic message templates
}
\details{
\code{read_template} and \code{write_template} provide basic input and output functionality for translation template (.pot) files. If called from with an R package directory, the locations of these fies are identified automatically (see \code{\link{make_po_dir}}).

\code{make_template} generates a new template in memory, without writing it to disk. \code{sync_template} makes a new template and writes it to disk or, if a template file already exists, overwrites it.
}
\examples{
# check for existing template
pkg <- extract_example_pkg()
try(check_for_template(pkg))

# generate an in-memory template
pot <- make_template(pkg)
write_template(pot)
  
# setup a package for translating
\dontrun{
# Should this be use_localization(pkg)?
use_translations(pkg)
}
}
\author{
Thomas J. Leeper
}
\seealso{
\code{\link{get_messages}} to read messages into memory without creating a template file, \code{\link{use_localization}} to setup a package for localization (including generation of a template file)
}

